+++ app/src/main/cpp/logger-master/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+v1.0.6
+
+- Added support for OpenBSD and other UNIX-like systems
+- Changed how shared pointers to parent logging objects are passed and stored
+- Changed the GetLogFacility() function to return the facility of the root
+  Logger object
+- Return empty string if GetCurrentTimestamp() produces an error
+- Removed #define for SYSLOG_SUPPORTED, relying on OS-specific #defines
+
+v1.0.5
+
+- Made compiler warnings stricter
+- Addressed warnings
+- Updated dependencies
+
+v1.0.4
+
+- Cosmetic changes
+- CMake changes
+
 v1.0.3
 
 - Changes required to build on OpenBSD